requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
google-generativeai
python-dotenv
PyPDF2
gTTS